[Onboarding] Separate popular artist into own component
